+++ crates/engine/src/plugin/wasm/error.rs
@@ -2,6 +2,7 @@ use crate::plugin::Func;
 use anyhow::Error;
 use displaydoc::Display;
 use std::io;
+use std::path::PathBuf;
 use thiserror::Error;
 use wasmtime::Trap;
 
@@ -16,49 +17,46 @@ pub enum Runtime {
 
     /// error running `{func}`
     Failed { func: Func, source: Trap },
+
+    /// invalid wasm module
+    InvalidModule(#[source] anyhow::Error),
+
+    /// unknown wasm error
+    Unknown(#[source] anyhow::Error),
+}
+
+impl From<anyhow::Error> for Runtime {
+    fn from(source: anyhow::Error) -> Self {
+        let cause = source.to_string();
+        if cause.contains("cross-`Store` instantiation is not currently supported") {
+            return Self::InvalidModule(source);
+        }
+        if cause.contains("wrong number of imports provided") {
+            return Self::InvalidModule(source);
+        }
+        if cause.starts_with("Bad") {
+            return Self::InvalidModule(source);
+        }
+        if cause.starts_with("expected") {
+            return Self::InvalidModule(source);
+        }
+
+        Self::Unknown(source)
+    }
 }
 
 /// `WasmPlugin` related errors.
 #[derive(Debug, Display, Error)]
 pub enum Handler {
     /// inaccessible wasm module `{path}` ({kind:?})
-    Io { path: String, kind: io::ErrorKind },
-
-    /// invalid wasm module `{path}`
-    InvalidModule { path: String, source: anyhow::Error },
-
-    /// error running wasm instance
-    Runtime(#[from] Runtime),
-
-    /// unknown wasm error for module `{path}`
-    Unknown { path: String, source: anyhow::Error },
+    Io { path: PathBuf, kind: io::ErrorKind },
 }
 
-impl From<(&str, anyhow::Error)> for Handler {
-    fn from((path, source): (&str, anyhow::Error)) -> Self {
-        let path = path.to_owned();
-
-        for cause in source.chain() {
-            if let Some(source) = cause.downcast_ref::<io::Error>() {
-                let kind = source.kind();
-                return Self::Io { path, kind };
-            }
+impl From<(PathBuf, io::Error)> for Handler {
+    fn from((path, err): (PathBuf, io::Error)) -> Self {
+        Self::Io {
+            path,
+            kind: err.kind(),
         }
-
-        let cause = source.to_string();
-        if cause.contains("cross-`Store` instantiation is not currently supported") {
-            return Self::InvalidModule { path, source };
-        }
-        if cause.contains("wrong number of imports provided") {
-            return Self::InvalidModule { path, source };
-        }
-        if cause.starts_with("Bad") {
-            return Self::InvalidModule { path, source };
-        }
-        if cause.starts_with("expected") {
-            return Self::InvalidModule { path, source };
-        }
-
-        Self::Unknown { path, source }
     }
 }