+++ crates/engine/src/plugin/wasm/plugin.rs
@@ -2,7 +2,7 @@ use super::RuntimeError;
 use crate::error;
 use crate::plugin::{Func, Runtime};
 use std::fmt;
-use wasmtime::Instance;
+use wasmtime::{Instance, Module, Store};
 
 /// A container type to wrap a Wasm module.
 pub struct Plugin {
@@ -11,8 +11,11 @@ pub struct Plugin {
 }
 
 impl Plugin {
-    pub(crate) const fn new(instance: Instance) -> Self {
-        Self { instance }
+    pub(super) fn new(store: &Store, source: impl AsRef<[u8]>) -> Result<Self, RuntimeError> {
+        let module = Module::new(store, source)?;
+        let instance = Instance::new(&module, &[])?;
+
+        Ok(Self { instance })
     }
 }
 
@@ -53,6 +56,36 @@ impl Runtime for Plugin {
 mod tests {
     use super::*;
 
+    mod new {
+        use super::*;
+
+        #[test]
+        fn valid() {
+            let wasm = r#"(module (func (export "_run")))"#;
+
+            assert!(plugin(wasm).is_ok())
+        }
+
+        #[test]
+        fn invalid_wasm() {
+            let wasm = "INVALID";
+
+            let err = anyhow::Error::new(plugin(wasm).unwrap_err());
+
+            assert_eq!(
+                format!("{:?}", err),
+                "invalid wasm module\n\n\
+
+                 Caused by:\n    \
+                     expected `(`\n         \
+                          --> <anon>:1:1\n          \
+                        |\n        \
+                      1 | INVALID\n          \
+                        | ^"
+            )
+        }
+    }
+
     mod run {
         use super::*;
 
@@ -60,14 +93,14 @@ mod tests {
         fn valid() {
             let wasm = r#"(module (func (export "_run")))"#;
 
-            assert!(plugin(wasm).run().is_ok())
+            assert!(plugin(wasm).expect("valid plugin").run().is_ok())
         }
 
         #[test]
         fn missing_function() {
             let wasm = r#"(module (func (export "_invalid")))"#;
 
-            let result = plugin(wasm).run();
+            let result = plugin(wasm).expect("valid plugin").run();
             let err = anyhow::Error::new(result.unwrap_err());
 
             assert_eq!(
@@ -83,7 +116,7 @@ mod tests {
         fn invalid_function_signature() {
             let wasm = r#"(module (func (export "_run") (result i32) i32.const 42))"#;
 
-            let result = plugin(wasm).run();
+            let result = plugin(wasm).expect("valid plugin").run();
             let err = anyhow::Error::new(result.unwrap_err());
 
             assert_eq!(
@@ -97,13 +130,8 @@ mod tests {
         }
     }
 
-    fn plugin(wasm: &str) -> Plugin {
-        use wasmtime::Module;
-
+    fn plugin(wasm: &str) -> Result<Plugin, RuntimeError> {
         let store = wasmtime::Store::default();
-        let module = Module::new(&store, wasm).unwrap();
-        let instance = Instance::new(&module, &[]).unwrap();
-
-        Plugin::new(instance)
+        Plugin::new(&store, wasm)
     }
 }